+++ dds_cli/motd_manager.py
@@ -58,7 +58,7 @@ class MotdManager(dds_cli.base.DDSBaseClass):
 
         # Only method "add"can use the MotdManager class
         if self.method not in ["add"]:
-            raise dds_cli.exceptions.AuthenticationError(f"Unauthorized method: '{self.method}'")
+            raise dds_cli.exceptions.InvalidMethodError(f"Unauthorized method: '{self.method}'")
 
     def add_new_motd(self, message):
         """Add a new motd."""
@@ -70,18 +70,63 @@ class MotdManager(dds_cli.base.DDSBaseClass):
             error_message="Failed adding a new MOTD",
         )
 
-        LOG.info("A new MOTD was added to the database")
+        response_message = response_json.get(
+            "message", "No response. Cannot confirm MOTD creation."
+        )
+        LOG.info(response_message)
 
     @staticmethod
-    def get_latest_motd():
-        """Get the latest MOTD from dabase"""
+    def list_all_active_motds(table=False):
+        """Get all active MOTDs."""
         try:
-            response_json, _ = dds_cli.utils.perform_request(
-                endpoint=DDSEndpoint.MOTD,
+            response, _ = dds_cli.utils.perform_request(
+                endpoint=dds_cli.DDSEndpoint.MOTD,
                 method="get",
-                error_message="Failed getting MOTD",
+                error_message="Failed getting MOTDs from API",
             )
         except:
             pass
         else:
-            return response_json.get("message")
+            # Get items from response
+            motd = response.get("motds")
+            if not motd:
+                message = response.get("message", "No motds or info message returned from API.")
+                LOG.info(message)
+            else:
+                motds, keys = dds_cli.utils.get_required_in_response(
+                    keys=["motds", "keys"], response=response
+                )
+                # Sort the active MOTDs according to date created
+                motds = dds_cli.utils.sort_items(items=motds, sort_by="Created")
+
+                # when called from "dds modt ls" with table=True
+                if table:
+                    # Create table
+                    table = dds_cli.utils.create_table(
+                        title="Active MOTDs.",
+                        columns=keys,
+                        rows=motds,
+                        ints_as_string=True,
+                        caption="Active MOTDs.",
+                    )
+
+                    # Print out table
+                    dds_cli.utils.print_or_page(item=table)
+                else:
+                    # on every dds call
+                    return motds
+
+    def deactivate_motd(self, motd_id):
+        """Deactivate specific MOTD."""
+        response_json, _ = dds_cli.utils.perform_request(
+            endpoint=DDSEndpoint.MOTD,
+            headers=self.token,
+            method="put",
+            json={"motd_id": motd_id},
+            error_message="Failed deactivating the MOTD",
+        )
+
+        response_message = response_json.get(
+            "message", "No response. Cannot confirm MOTD deactivation."
+        )
+        LOG.info(response_message)